Move cli/util.go to cli/util/flags.go
